+++ src/main/java/com/chpark/calendar/repository/schedule/ScheduleRepeatRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ScheduleRepeatRepository extends JpaRepository<ScheduleRepeatEntity, Integer> {
+public interface ScheduleRepeatRepository extends JpaRepository<ScheduleRepeatEntity, Long> {
 
 }
